requirements.txt
gradio
numpy
pandas
GoogleNews
newspaper3k
nltk